Refactor AI Generator: extract reusable components
Moved AnimatedMeshGradient, RecipeTagSection, the static unavailable card
(AIUnavailableCard), and the pressEvents modifier into AIGeneratorComponents.swift.
AIRecipeGeneratorView drops from 852 to ~700 lines and now focuses on the
generation flow/state; the deeply-stateful input/preview sections stay put.
