+++ Cauldron/Features/AIGenerator/AIRecipeGeneratorView.swift
@@ -30,7 +30,7 @@ struct AIRecipeGeneratorView: View {
                 ScrollView {
                     VStack(spacing: Theme.Spacing.xl) {
                         if !isAvailable {
-                            unavailableSection
+                            AIUnavailableCard()
                         } else {
                             // Combined section that expands/collapses
                             expandableInputSection
@@ -432,24 +432,6 @@ struct AIRecipeGeneratorView: View {
         .cornerRadius(12)
     }
 
-    private var unavailableSection: some View {
-        VStack(spacing: 16) {
-            Image(systemName: "exclamationmark.triangle")
-                .font(.system(size: 44))
-                .foregroundColor(.orange)
-
-            Text("Apple Intelligence Not Available")
-                .font(.headline)
-
-            Text("This device doesn't support Apple Intelligence, or it may be disabled in Settings.")
-                .font(.body)
-                .foregroundColor(.secondary)
-                .multilineTextAlignment(.center)
-        }
-        .padding(24)
-        .cardStyle()
-    }
-
     private var combinedInputSection: some View {
         VStack(alignment: .leading, spacing: 20) {
             // Prompt/Notes field
@@ -754,98 +736,6 @@ struct AIRecipeGeneratorView: View {
 }
 
 // MARK: - Recipe Tag Section
-
-struct RecipeTagSection: View {
-    let title: String
-    let icon: String
-    let tags: [RecipeCategory]
-    let onTagTap: (RecipeCategory) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            Label(title, systemImage: icon)
-                .font(.caption)
-                .fontWeight(.bold)
-                .foregroundColor(.secondary)
-                .textCase(.uppercase)
-                .padding(.horizontal, 4)
-
-            ScrollView(.horizontal, showsIndicators: false) {
-                HStack(spacing: 10) {
-                    ForEach(tags, id: \.self) { tag in
-                        Button {
-                            onTagTap(tag)
-                        } label: {
-                            TagView(tag.tagValue)
-                        }
-                        .buttonStyle(.plain)
-                    }
-                }
-                .padding(.horizontal, 4)
-                .padding(.vertical, 4) // Space for shadow
-            }
-        }
-    }
-}
-
-
-
-// MARK: - Animated Mesh Gradient
-
-struct AnimatedMeshGradient: View {
-    @State private var animate = false
-    
-    var body: some View {
-        ZStack {
-            // Base background
-            Color.cauldronBackground
-            
-            // Moving orbs
-            GeometryReader { proxy in
-                ZStack {
-                    // Orb 1
-                    Circle()
-                        .fill(Color.orange.opacity(0.15))
-                        .frame(width: proxy.size.width * 0.8)
-                        .blur(radius: 60)
-                        .offset(x: animate ? -50 : 50, y: animate ? -50 : 50)
-                    
-                    // Orb 2
-                    Circle()
-                        .fill(Color.yellow.opacity(0.15))
-                        .frame(width: proxy.size.width * 0.7)
-                        .blur(radius: 60)
-                        .offset(x: animate ? 100 : -100, y: animate ? 100 : -50)
-                    
-                    // Orb 3
-                    Circle()
-                        .fill(Color.red.opacity(0.1))
-                        .frame(width: proxy.size.width * 0.6)
-                        .blur(radius: 50)
-                        .offset(x: animate ? -30 : 150, y: animate ? 150 : -30)
-                }
-                .frame(maxWidth: .infinity, maxHeight: .infinity)
-            }
-        }
-        .onAppear {
-            withAnimation(.easeInOut(duration: 10).repeatForever(autoreverses: true)) {
-                animate.toggle()
-            }
-        }
-    }
-}
-
-// MARK: - View Modifiers
-
-extension View {
-    func pressEvents(onPress: @escaping (Bool) -> Void) -> some View {
-        self.simultaneousGesture(
-            DragGesture(minimumDistance: 0)
-                .onChanged { _ in onPress(true) }
-                .onEnded { _ in onPress(false) }
-        )
-    }
-}
 
 #Preview {
     AIRecipeGeneratorView(dependencies: .preview())
